+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Models/VRChat/AvatarConverter.cs
@@ -204,11 +204,12 @@ namespace KRT.VRCQuestTools.Models.VRChat
             questAvatarObject.SetActive(true);
 
             var converterSetttigs = questAvatarObject.GetComponent<Components.AvatarConverterSettings>();
-#if VQT_HAS_VRCSDK_LOCAL_CONTACT_RECEIVER
-            var contactsToKeep = converterSetttigs.contactsToKeep.Concat(avatar.GetLocalContactReceivers()).Distinct().ToArray();
-#else
-            var contactsToKeep = converterSetttigs.contactsToKeep;
-#endif
+            var contactsToKeep = converterSetttigs.contactsToKeep
+                .Concat(avatar.GetLocalContactReceivers())
+                .Concat(avatar.GetLocalContactSenders())
+                .Distinct()
+                .ToArray();
+
             VRCSDKUtility.DeleteAvatarDynamicsComponents(avatar, converterSetttigs.physBonesToKeep, converterSetttigs.physBoneCollidersToKeep, contactsToKeep);
 
             UnityEngine.Object.DestroyImmediate(converterSetttigs);
@@ -322,7 +323,9 @@ namespace KRT.VRCQuestTools.Models.VRChat
 
         private void RemoveExtraMaterialSlots(GameObject questAvatarObject)
         {
-            var renderers = questAvatarObject.GetComponentsInChildren<Renderer>(true);
+            var renderers = questAvatarObject.GetComponentsInChildren<SkinnedMeshRenderer>(true)
+                .Cast<Renderer>()
+                .Concat(questAvatarObject.GetComponentsInChildren<MeshRenderer>(true));
             foreach (var renderer in renderers)
             {
                 var mesh = RendererUtility.GetSharedMesh(renderer);