+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/NDMF/Common/NdmfPluginUtility.cs
@@ -49,6 +49,8 @@ namespace KRT.VRCQuestTools.Ndmf
                 return;
             }
 
+            context.GetState<NdmfState>().compressExpressionsMenuIcons = settings.compressExpressionsMenuIcons;
+
             try
             {
                 TrackObjectRegistryForMaterialSwaps(context.AvatarRootObject);