+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Utils/VRCSDKUtility.cs
@@ -537,6 +537,28 @@ namespace KRT.VRCQuestTools.Utils
         }
 
         /// <summary>
+        /// Gets the contact is local-only.
+        /// </summary>
+        /// <param name="contact">Contact to inspect.</param>
+        /// <returns>true when the contact is local-only.</returns>
+        internal static bool IsLocalOnlyContact(ContactBase contact)
+        {
+#if VQT_HAS_VRCSDK_LOCAL_CONTACT_RECEIVER
+            if (contact is ContactReceiver receiver)
+            {
+                return receiver.IsLocalOnly;
+            }
+#endif
+#if VQT_HAS_VRCSDK_LOCAL_CONTACT_SENDER
+            if (contact is ContactSender sender)
+            {
+                return sender.IsLocalOnly;
+            }
+#endif
+            return false;
+        }
+
+        /// <summary>
         /// Stripes unused network ids from the avatar.
         /// </summary>
         /// <param name="avatarDescriptor">Target avatar.</param>
@@ -707,6 +729,99 @@ namespace KRT.VRCQuestTools.Utils
             var knownMenu = new HashSet<VRCExpressionsMenu>();
             GetMenuTexturesFromMenuImpl(menu, textures, knownMenu);
             return textures.ToArray();
+        }
+
+        /// <summary>
+        /// Duplicates the expressions menu.
+        /// </summary>
+        /// <param name="rootMenu">Root menu to duplicate.</param>
+        /// <returns>Duplicated menu.</returns>
+        internal static VRCExpressionsMenu DuplicateExpressionsMenu(VRCExpressionsMenu rootMenu)
+        {
+            if (rootMenu == null)
+            {
+                return null;
+            }
+
+            var menuMap = new Dictionary<VRCExpressionsMenu, VRCExpressionsMenu>();
+            VRCExpressionsMenu Duplicate(VRCExpressionsMenu menu)
+            {
+                var newMenu = ScriptableObject.Instantiate(menu);
+                newMenu.name = menu.name + " (VQT Clone)";
+                menuMap.Add(menu, newMenu);
+                for (var i = 0; i < menu.controls.Count; i++)
+                {
+                    var subMenu = menu.controls[i].subMenu;
+                    if (subMenu != null)
+                    {
+                        if (menuMap.ContainsKey(subMenu))
+                        {
+                            newMenu.controls[i].subMenu = menuMap[subMenu];
+                        }
+                        else
+                        {
+                            newMenu.controls[i].subMenu = Duplicate(subMenu);
+                        }
+                    }
+                }
+                return newMenu;
+            }
+
+            return Duplicate(rootMenu);
+        }
+
+        /// <summary>
+        /// Resizes the icons in the expressions menu.
+        /// </summary>
+        /// <param name="rootMenu">Root menu.</param>
+        /// <param name="maxSize">Max texture size. Set 0 to remove.</param>
+        /// <param name="compressTextures">Whether to compress textures. Compress them in progressCallback.</param>
+        /// <param name="progressCallback">Callback for created textures.</param>
+        internal static void ResizeExpressionMenuIcons(VRCExpressionsMenu rootMenu, int maxSize, bool compressTextures, Action<Texture2D, Texture2D> progressCallback)
+        {
+            if (rootMenu == null)
+            {
+                return;
+            }
+
+            HashSet<VRCExpressionsMenu> knownMenus = new HashSet<VRCExpressionsMenu>();
+            Dictionary<Texture2D, Texture2D> resizedTextures = new Dictionary<Texture2D, Texture2D>();
+            void ResizeExpressionMenuIconsImpl(VRCExpressionsMenu menu)
+            {
+                knownMenus.Add(menu);
+                foreach (var control in menu.controls)
+                {
+                    if (maxSize == 0)
+                    {
+                        control.icon = null;
+                    }
+                    else if (control.icon != null)
+                    {
+                        var icon = control.icon;
+                        var needToCompress = compressTextures && AssetUtility.IsUncompressedFormat(icon.format);
+                        if (resizedTextures.ContainsKey(icon))
+                        {
+                            control.icon = resizedTextures[icon];
+                        }
+                        else if (icon.width > maxSize || icon.height > maxSize || needToCompress)
+                        {
+                            var newWidth = Math.Min(maxSize, icon.width);
+                            var newHeight = Math.Min(maxSize, icon.height);
+                            var newIcon = AssetUtility.ResizeTexture(icon, newWidth, newHeight);
+                            newIcon.name = icon.name + " (VQT Resize)";
+                            control.icon = newIcon;
+                            resizedTextures.Add(icon, newIcon);
+                            progressCallback?.Invoke(icon, newIcon);
+                        }
+                    }
+
+                    if (control.subMenu != null && !knownMenus.Contains(control.subMenu))
+                    {
+                        ResizeExpressionMenuIconsImpl(control.subMenu);
+                    }
+                }
+            }
+            ResizeExpressionMenuIconsImpl(rootMenu);
         }
 
         private static void GetMenuTexturesFromMenuImpl(VRCExpressionsMenu menu, HashSet<Texture2D> textures, HashSet<VRCExpressionsMenu> knownMenus)
@@ -909,22 +1024,20 @@ namespace KRT.VRCQuestTools.Utils
                 /// </summary>
                 internal Transform RootTransform => (Transform)RootTransformField.GetValue(component);
 
-#if VQT_HAS_VRCSDK_LOCAL_CONTACT_RECEIVER
                 /// <summary>
-                /// Gets whether the component is local only.
+                /// Gets a value indicating whether the component is local only.
                 /// </summary>
                 internal bool IsLocalOnly
                 {
                     get
                     {
-                        if (component is ContactReceiver receiver)
+                        if (component is VRC.Dynamics.ContactBase contact)
                         {
-                            return receiver.IsLocalOnly;
+                            return IsLocalOnlyContact(contact);
                         }
                         return false;
                     }
                 }
-#endif
             }
         }
     }