+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/ViewModels/MSMapGenViewModel.cs
@@ -7,6 +7,7 @@ using System;
 using KRT.VRCQuestTools.Utils;
 using UnityEditor;
 using UnityEngine;
+using UnityEngine.Rendering;
 
 namespace KRT.VRCQuestTools.ViewModels
 {
@@ -49,26 +50,14 @@ namespace KRT.VRCQuestTools.ViewModels
                 var width = Math.Max(metallic.Object?.width ?? 4, smoothness.Object?.width ?? 4);
                 var height = Math.Max(metallic.Object?.height ?? 4, smoothness.Object?.height ?? 4);
 
-                using (var dstTexture = DisposableObject.New(new RenderTexture(width, height, 0, RenderTextureFormat.ARGB32, RenderTextureReadWrite.Linear)))
-                {
-                    baker.Object.SetTexture("_MetallicMap", metallic.Object);
-                    baker.Object.SetInt("_InvertMetallic", 0);
-                    baker.Object.SetTexture("_SmoothnessMap", smoothness.Object);
-                    baker.Object.SetInt("_InvertSmoothness", invertSmoothness ? 1 : 0);
+                var dstTexture = new RenderTexture(width, height, 0, RenderTextureFormat.ARGB32, RenderTextureReadWrite.Linear);
+                baker.Object.SetTexture("_MetallicMap", metallic.Object);
+                baker.Object.SetInt("_InvertMetallic", 0);
+                baker.Object.SetTexture("_SmoothnessMap", smoothness.Object);
+                baker.Object.SetInt("_InvertSmoothness", invertSmoothness ? 1 : 0);
 
-                    // Remember active render texture
-                    var activeRenderTexture = RenderTexture.active;
-                    Graphics.Blit(metallic.Object, dstTexture.Object, baker.Object);
-
-                    Texture2D outTexture = new Texture2D(width, height);
-                    outTexture.ReadPixels(new Rect(0, 0, width, height), 0, 0);
-                    outTexture.Apply();
-
-                    // Restore active render texture
-                    RenderTexture.active = activeRenderTexture;
-
-                    AssetUtility.SaveUncompressedTexture(destPath, outTexture, false);
-                }
+                var outTexture = AssetUtility.BakeTexture(metallic.Object, baker.Object, width, height, false);
+                AssetUtility.SaveUncompressedTexture(destPath, outTexture, false);
             }
         }
     }