+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Views/PhysBonesRemoveWindow.cs
@@ -169,7 +169,9 @@ namespace KRT.VRCQuestTools.Views
                 }
                 EditorGUILayout.EndFoldoutHeaderGroup();
 
-#if VQT_HAS_VRCSDK_LOCAL_CONTACT_RECEIVER
+#if VQT_HAS_VRCSDK_LOCAL_CONTACT_SENDER
+                var contactsHeader = "Non-Local Contact Senders & Non-Local Contact Receivers";
+#elif VQT_HAS_VRCSDK_LOCAL_CONTACT_RECEIVER
                 var contactsHeader = "Contact Senders & Non-Local Contact Receivers";
 #else
                 var contactsHeader = "Contact Senders & Contact Receivers";
@@ -178,11 +180,7 @@ namespace KRT.VRCQuestTools.Views
                 {
                     using (var vertical = new EditorGUILayout.VerticalScope(foldedContentPanel))
                     {
-#if VQT_HAS_VRCSDK_LOCAL_CONTACT_RECEIVER
                         var contacts = model.Avatar.GetNonLocalContacts();
-#else
-                        var contacts = model.Avatar.GetContacts();
-#endif
                         if (contacts.Length > 0)
                         {
                             using (var horizontal = new EditorGUILayout.HorizontalScope())